requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.2
django-rest-auth==0.9.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
PyJWT==2.8.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
urllib3==2.2.1
